Make PayPalRecurringButton work for subscriptions as well as contributions
