+++ support-frontend/assets/helpers/paymentIntegrations/payPalActions.js
@@ -1,0 +1,5 @@
+// @flow
+
+export type Action = { type: 'SET_PAYPAL_HAS_LOADED' };
+
+export const setPayPalHasLoaded = (): Action => ({ type: 'SET_PAYPAL_HAS_LOADED' });
